Fix compiler errors.
git-svn-id: https://svn.openfabrics.org/svnrepo/nvmewin/releases/revision_1.5@144 5815846e-57a8-46bc-a7d2-e79180fb8327

+++ source/nvmeSnti.c
@@ -9010,12 +9010,12 @@ VOID SntiDpcRoutine(
 {
     PNVME_SRB_EXTENSION pSrbExt = NULL;
     PNVME_DEVICE_EXTENSION pDevExt = NULL;
-    UINT32 bufferSize;
+    SIZE_T bufferSize;
     ULONG status;
 
     pDevExt = (PNVME_DEVICE_EXTENSION)pHwDeviceExtension;
     pSrbExt = (PNVME_SRB_EXTENSION)pSystemArgument1;
-    bufferSize = (UINT32)pSystemArgument2;
+    bufferSize = (SIZE_T)pSystemArgument2;
 
     status = StorPortFreeContiguousMemorySpecifyCache(pDevExt,
                                                       pSrbExt->pDataBuffer,
